requirements.txt
Flask==2.1.0
python-docx==0.8.11
gunicorn==20.1.0
google-cloud-tasks==2.7.1
